requirements.txt
fastapi==0.103.0
uvicorn==0.23.0
sqlalchemy==2.0.0
asyncpg==0.27.0
psycopg2==2.9.6
alembic==1.10.0
pydantic==1.10.7
python-dotenv==0.21.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
requests==2.28.1
